requirements.txt
# Core dependencies
langgraph==0.2.56
langchain==0.3.11
langchain-community==0.3.10
langchain-openai==0.2.13

# Azure OpenAI
openai==1.57.0

# Azure Document Intelligence
azure-ai-formrecognizer==3.3.3

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.11.4
pymupdf==1.24.14
reportlab==4.2.5

# Data processing
pydantic==2.10.4
python-dotenv==1.0.1

# Additional utilities
typing-extensions==4.12.2